Extract Self formatter
