+++ lib/awesome_print/formatters.rb
@@ -1,6 +1,7 @@
 module AwesomePrint
   module Formatters
     require 'awesome_print/formatters/base'
+    require 'awesome_print/formatters/self'
     require 'awesome_print/formatters/array'
     require 'awesome_print/formatters/hash'
     require 'awesome_print/formatters/object'
